jobs/ml-ops/container/inference-api/requirements.txt
fastapi==0.104.1
boto3==1.29.3
uvicorn==0.24.0.post1
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0